Revert "In progress"
This reverts commit 9f0891d2f9b7749d0c0de89045b8b1220130f4b1.

+++ TestSutraUsers/src/main/java/com/tathvatech/user/OID/OID.java
@@ -59,7 +59,7 @@ public abstract class OID extends TSBeanBase implements Authorizable{
 	}
 
 	@Override
-	public int getPk() {
+	public long getPk() {
 		return (int) pk;
 	}
